+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.5.0] - 2022-11-18
+
+### Added
+
+- Added support for internal command "_uid" (Kilovolt protocol v9)
+
+### Fixed
+
+- Version command was incorrectly specified as "kversion" instead of "version"
+
 ## [6.4.0] - 2022-02-23
 
 ### Added
@@ -32,6 +42,7 @@ Rewrote the entire package to be Deno-compatible and Deno-first.
 
 - Added v6 compatibility, changed constructor to accept a password parameter for authentication
 
+[6.5.0]: https://github.com/strimertul/kilovolt-client-ts/compare/v6.4.0...v6.5.0
 [6.4.0]: https://github.com/strimertul/kilovolt-client-ts/compare/v6.3.2...v6.4.0
 [6.3.2]: https://github.com/strimertul/kilovolt-client-ts/compare/v6.3.1...v6.3.2
 [6.3.1]: https://github.com/strimertul/kilovolt-client-ts/compare/v6.0.0...v6.3.1